requirements.txt
fastapi
uvicorn
python-multipart==0.0.5
mysql-connector-python

